romiseg/pyproject.toml
[build-system]
requires = ["setuptools >=75"]
build-backend = "setuptools.build_meta"

[project]
name = "romiseg"
version = "0.2.0"
dependencies = [
    "appdirs",
    "future",
    "Pillow",
    "requests",
    "torch<2",
    "torchvision",
    "tqdm",
    "plantdb.commons",
]
description = "Image multiclass semantic segmentation using CNN models trained on virtual plant images (PyTorch)."
readme = "README.md"
requires-python = ">=3.8"
license = { 'text' = "LGPL-3.0-or-later" }   # to replace by "LGPL-3.0-or-later" only in next release, pending deprecation
authors = [
    { name = "Alienor Lahlou", email = "alienor.lahlou@espci.org" },
    { name = "Peter Hanappe", email = "peter.hanappe@sony.com" },
]
maintainers = [
    { name = "Jonathan Legrand", email = "jonathan.legrand@ens-lyon.fr" }
]
keywords = [
    "Robotics for Microfarms",
    "ROMI",
    "U-Net",
    "Deep CNN",
    "Virtual Plant",
    "semantic segmentation",
    "PyTorch",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
    "Natural Language :: English",
]

[project.scripts]
finetune = "romiseg.finetune:main"
train_cnn = "romiseg.train_cnn:main"

[project.urls]
homepage = "https://romi-project.eu/"
documentation = "https://docs.romi-project.eu/plant_imager/"
source = "https://github.com/romi/romiseg"
issues = "https://github.com/romi/romiseg/issues"

[project.optional-dependencies]
train = [
    "labelme",
    "pyqt5==5.14",
    "tensorboard",
]
